Rename refresh saga to sync saga.
